Added some html files to layouts partials to override template
I have also corrected some author typos in a bunch of publications.
Changed the params.yaml file to add an option to limit the number of
publications that appear in the latest section under author profiles. I
have also modifies the layouts/partials/author/list.html file so as it
uses this parameter.