Added a block in layouts blocks about.hml for author selected  papers
I have added a block in layouts/partials/blocks/v1/about.html so as it
adds a 100% width space below the interests and education sections. This
new section is conditional to the existence of front matter
"selected_papers" in theauthor's profile file. It is a mrkdown section
really, but it is intended to add selected references chosen by the
author. Front matter looks like this:
custom_markdown_section: |
  - Massana et al., 2024 .....