Add filter buttons in author profile page
I have modified the author page to show buttons that point to the
publications page, filtered by author name. To do that I have had to
modify the javascript file for publications and to add an extra button
in the layouts/section/publications.html page. At the moment, this file
hold a list of hard coded names and slugs for the filter to be active. I
have also added custom CSS style for the buttons in author profile pages
and citation view, using the theme colors.