Add author name and slug list to params.yaml
I did not like the hard coded names in the publication.html file, so I
modified the filtering code to parse a new author list in params.yaml.
This list has the author names and slugs, so as the query URL can be
crafted properly. I have had to remove the dot in the middle
name/initial of some authors. This is affecting to the paper assignation
to each author and has to do with the initial having a dot in the author
names in the front matter of each publication.

I need to address this in the index.html files of each publication.