Added filtering by thesis supervisor in URL
I have modified the publication.html layout and the javascript that
filters publications so as now I have an additional filter that acts on
publication-type thesis and uses the list of EMM members as query for
the front matter 'supervisor' that should be in each thesis record in
content/publication. I have added a new scss class to hide the dropdown
menu. As a note, filtering supervisors via URL works beautifully, but it
didn't work when using the dropdown menu for some authors.