Add missing EMM members
I have added several PhD students, technicians and postdocs that were
not in the website. I have added a barebones page, with only the name
and the email. I have added a weight parameter for each author to fix
the sorting issue. It is a manual sorting strategy, but it works.
I have also added a couple of former members