Fix button links. Now this is the fix
I have fixed the links in the author profile buttons. The thing is that
locally I have to add the project name and now that we have added a
personalized domain to the github.io site, all the urls that had the
pathe /emm/ or /emmTest/ now don't work. I have ahd to remove that path
MAKING SURE that the new paths start at the root of the site with a '/'