Tackle the problem with periods in author names
It is making a lot of confussion to have authors with middle name
initials with and without period. I have made here a couple of tests
with pep's profile and it seems to work. The idea is to modify the slug
in params.yaml to read as 'josep-m\.-gasol' and this makes the filter
dropdown menu to work fine. Also, in the author.html layout, the buttons
have been coded so as they modify the URL to escape the period. I know
that this may be a capital sin, and dots should be coded like %2E or
something, but the browser just keeps converting the ASCII code in a
literl '.', breaking down the functionality.