Added Bluesky icon
I have added a custom icon for bluesky in static/images. Then I have
added an scss class for custom icons and I have edited the about.html
template to check whether there's a custom icon or else use the font
awesome set. Itried to update the font awesome set that is pulled for
the template, but I couldn't so this is a next-best-fix